+++ kbe/src/server/dbmgr/dbmgr.cpp
@@ -850,7 +850,7 @@ void Dbmgr::executeRawDatabaseCommand(Network::Channel* pChannel,
 //-------------------------------------------------------------------------------------
 PyObject* Dbmgr::__py_executeRawDatabaseCommand(PyObject* self, PyObject* args)
 {
-	int argCount = (int)PyTuple_Size(args);
+	Py_ssize_t argCount = PyTuple_Size(args);
 	PyObject* pycallback = NULL;
 	PyObject* pyDBInterfaceName = NULL;
 	int ret = 0;
@@ -1169,7 +1169,7 @@ void Dbmgr::queryEntity(Network::Channel* pChannel, uint16 dbInterfaceIndex, COM
 //-------------------------------------------------------------------------------------
 void Dbmgr::syncEntityStreamTemplate(Network::Channel* pChannel, KBEngine::MemoryStream& s)
 {
-	int rpos = s.rpos();
+	size_t rpos = s.rpos();
 	EntityTables::ENTITY_TABLES_MAP::iterator iter = EntityTables::sEntityTables.begin();
 	for (; iter != EntityTables::sEntityTables.end(); ++iter)
 	{
